refactor: separate ui-renderer files for react-native
